+++ src/layouts/AuthLayout.vue
@@ -12,7 +12,7 @@
           <va-tabs v-model="tabIndex" center>
             <template #tabs>
               <va-tab name="login">Login</va-tab>
-              <va-tab name="signup">Create New Account</va-tab>
+              <!--<va-tab name="signup">Create New Account</va-tab>-->
             </template>
           </va-tabs>
