+++ src/router/index.ts
@@ -223,6 +223,16 @@ const routes: Array<RouteRecordRaw> = [
         component: () => import('../pages/auth/recover-password/RecoverPassword.vue'),
       },
       {
+        name: 'reset-password',
+        path: 'reset-password',
+        component: () => import('../pages/auth/reset-password/ResetPassword.vue'),
+      },
+      {
+        name: 'first-time',
+        path: 'first-time',
+        component: () => import('../pages/auth/login/FirstTimeUser.vue'),
+      },
+      {
         path: '',
         redirect: { name: 'login' },
       },
